LUGG-529 Consistency in Content Types
Added the location field for convenience's sake.

Rearranged fields so that the location field was located before the body text but above the event type, as it is on the NREM site.

Updated field settings to be more in line with other content types (like image and file settings, for example). Also added a description for tags, which is what other content types have.